nit: code-review fix
